moved AbstractTokenizer into own file, made CmdTokenizer its child
